sdm: ensure color mode stays the same after reboot
ColorIntfGetActiveMode() may return -1 when the display is not
ready during boot. Store the value of color_mode instead, so
the display is not always set to sRGB after reboot.

Ignore the sRGB transform from frameworks when color_mode_before_p3_
is uninitialized, so that the previous sRGB profile can be retained
after reboot. This is not an issue when the display only has a
single sRGB color profile, but when the display has more than one,
the color mode will always be set to a specific sRGB profile,
regardless of the settings before reboot.

Test: check both DCI P3 profile and sRGB profile are retained
      after reboot
Test: check sRGB_d65 does not change to srgb_native after reboot

Change-Id: Ifb905a5c8d939a24d0a7f38363e2415c0e6ef03f
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: ALKALiKong <alikongalk@gmail.com>